refactor(avatar): make the character catalog one shared package
The catalog had three readers and three mechanisms: the daemon imported it
locally, the iOS icon scripts reached into assistant/src by relative path,
and clients/web kept a 437-line hand copy carrying a keep-in-sync note. The
copy was byte-identical today, so nothing had drifted yet, but nothing
stopped it either.

It now lives in @vellumai/avatar-catalog, a leaf package with no
dependencies, and every reader imports it. The web copy becomes a
nineteen-line adapter that exposes the catalog as a module-scope value,
which is what its consumers need: getCharacterComponents() builds a fresh
object per call, and web code passes the result into React effect
dependencies.

The package keeps the palette in its own module behind a ./colors subpath
export, so boot-path code that resolves a stored colour id does not pull the
47 kB of SVG path data with it. Verified on a production build: no body
shape reaches the entry chunk, the payload stays in its own lazy chunk, and
the extracted catalog is byte-for-byte the same 46,134 characters the daemon
served before.
