+++ clients/ios/scripts/__tests__/android-launcher-geometry.test.ts
@@ -12,7 +12,7 @@ import { describe, expect, test } from "bun:test";
 import { readFileSync } from "node:fs";
 import { join } from "node:path";
 
-import { getCharacterComponents } from "../../../../assistant/src/avatar/character-components.js";
+import { getCharacterComponents } from "../../../../packages/avatar-catalog/src/index.js";
 
 const ANDROID_RES = join(import.meta.dir, "../../../android/app/src/main/res");
