+++ clients/web/src/utils/avatar-bundled-components.ts
@@ -1,433 +1,19 @@
 /**
- * Static bundled copy of the character-components data so the hatching screen
- * can compose avatars without needing the daemon to be running.
+ * The character catalog as a module-scope value, so the app can compose an
+ * avatar before the daemon answers `/avatar/character-components` (the
+ * hatching screen, the assistant chooser, a transcript's subagent chips).
  *
- * Canonical source: assistant/src/avatar/character-components.ts
- * Keep in sync when body shapes, eye styles, or colors change.
+ * A value rather than a call at each site: consumers pass this straight into
+ * React components and effect dependencies, and `getCharacterComponents()`
+ * builds a fresh object every call, so calling per site would hand them a new
+ * identity on every render.
+ *
+ * This module carries the ~47 kB of SVG path data. Boot-path code that needs
+ * only the palette imports `@/utils/avatar-bundled-colors` instead.
  */
 
+import { getCharacterComponents } from "@vellumai/avatar-catalog";
+
 import type { CharacterComponents } from "@/types/avatar";
-import { BUNDLED_COLORS } from "@/utils/avatar-bundled-colors";
 
-const SCLERA = "#F2F2F2";
-const PUPIL = "#1A1A1A";
-
-export const BUNDLED_COMPONENTS: CharacterComponents = {
-  bodyShapes: [
-    {
-      id: "blob",
-      viewBox: { width: 580, height: 561 },
-      faceCenter: { x: 300, y: 273 },
-      svgPath:
-        "M551.3 377.649C524.596 440.214 479.223 495.75 419.697 527.13C383.966 545.972 344.481 554.698 304.742 558.505C263.739 562.448 221.967 561.741 181.567 552.87C112.437 537.704 44.9966 494.446 17.5938 424.615C6.00081 395.096 2.48816 362.476 3.26999 330.427C2.55845 325.429 1.96158 320.43 1.46585 315.418C-2.60607 273.686 1.79634 231.456 14.7865 191.69C39.2421 116.861 93.1524 51.6488 165.738 23.6091C206.151 8.00384 250.571 1.15769 293.581 0.10974C336.598 -0.938208 379.647 5.52128 420.516 19.4032C457.676 32.0143 494.76 50.2863 522.595 79.2557C546.619 104.249 562.372 136.396 571.03 170.227C588.555 238.703 578.729 313.388 551.3 377.649Z",
-    },
-    {
-      id: "cloud",
-      viewBox: { width: 700, height: 639 },
-      faceCenter: { x: 350, y: 380 },
-      svgPath:
-        "M458.043 529.146C461.866 530.414 451.569 563.252 450.19 566.713C445.375 578.789 438.393 589.856 430.133 599.861C397.68 639.164 349.367 647.078 302.344 631.557C267.032 619.903 245.62 595.757 234.464 560.446C233.64 557.836 232.873 555.207 231.85 551.831C215.336 566.821 196.672 575.311 175.804 578.771C159.74 581.434 143.674 580.153 127.916 576.427C122.088 575.05 116.003 572.978 111.13 569.631C59.452 534.135 37.9953 455.647 82.0824 405.11C84.6116 402.21 87.0818 399.258 90.4356 395.327C87.0013 394.381 84.8031 393.586 82.5348 393.182C61.0136 389.341 43.0259 378.373 27.2162 364.12C-1.3201 338.393 -7.78894 301.77 9.57619 267.482C22.6394 241.688 40.3066 220.148 67.513 208.004C87.5677 199.052 108.768 196.873 130.414 197.514C135.727 197.672 141.026 198.286 148.39 198.846C145.862 194.159 144.099 190.889 142.336 187.62C124.709 154.933 122.252 121.389 135.068 86.2933C154.162 34.0028 208.103 6.92757 261.126 18.1841C292.089 24.7577 327.264 42.3061 342.454 72.2215C343.254 73.7957 344.247 75.2755 345.215 76.7552C345.351 76.9621 345.865 76.9211 346.994 77.1668C348.274 74.2177 349.567 71.0638 350.999 67.9745C355.052 59.2326 358.588 50.1887 363.355 41.8501C381.347 10.3757 420.307 -5.42286 455.406 1.68033C464.854 3.59251 472.777 8.07002 479.606 14.6749C498.419 32.8736 504.856 55.5692 503.566 81.0162C503.229 87.6502 502.673 94.2724 502.091 102.759C506.659 99.8181 509.529 97.9475 512.422 96.1117C538.175 79.7752 566.338 75.0749 595.791 81.2105C630.232 88.3845 655.025 116.808 661.425 152.356C665.943 177.46 661.851 201.584 654.716 225.279C652.268 233.41 645.875 240.356 641.318 247.788C649.735 253.395 658.76 259.005 667.358 265.211C691.529 282.657 701.903 310.011 699.081 339.212C695.999 371.1 679.845 396.926 647.129 404.045C633.529 407.004 619.694 409.116 605.645 406.53C603.304 406.099 600.817 406.469 596.527 406.469C599.85 411.595 602.3 415.803 605.159 419.713C618.729 438.28 621.763 460.019 621.884 481.845C621.95 493.691 618.891 506.205 611.341 516.237C604.493 525.336 597.416 534.524 589.068 542.178C574.463 555.569 555.851 558.53 536.924 557.352C509.254 555.63 483.751 546.931 461.278 530.245C460.75 529.856 459.986 529.791 458.043 529.146Z",
-    },
-    {
-      id: "sprout",
-      viewBox: { width: 613, height: 628 },
-      faceCenter: { x: 264, y: 415 },
-      svgPath:
-        "M188.562 154.32C195.642 129.19 201.743 104.963 209.452 81.259C213.281 69.4883 219.31 58.2302 225.661 47.5184C241.742 20.3937 274.883 -0.450409 307.341 0.0073979C347.653 0.57583 371.525 35.5013 378.375 71.9457C382.152 92.0349 381.242 112.022 377.249 131.965C376.161 137.396 375.235 142.859 374.295 147.987C386.293 137.766 398.035 127.685 409.861 117.703C438.964 93.1439 473.371 80.9459 510.384 75.0849C550.196 68.7806 620 71.8795 612.432 129.169C609.445 151.779 597.67 170.563 583.001 187.388C563.368 209.908 539.237 227.292 515.201 244.7C508.119 249.829 500.508 254.317 493.876 259.962C466.199 283.521 448.892 313.117 444.204 349.376C443.415 355.485 444.351 362.451 446.605 368.192C458.814 399.293 478.634 429.571 480.806 463.723C484.301 518.658 460.669 566.367 412.455 593.049C351.18 626.96 277.948 634.454 209.647 622.315C153.874 612.402 85.8161 587.91 68.0917 527.969C51.6962 472.524 73.4845 423.106 86.6231 369.772C91.1017 351.591 89.5279 333.925 80.9861 316.585C69.3865 293.036 57.5037 269.791 42.8657 247.887C31.4803 230.851 21.2917 212.963 11.3918 195.003C0.986137 176.125 -2.15173 155.792 1.40914 134.331C9.90013 83.1647 65.013 77.6815 104.347 95.9784C115.372 101.107 126.539 105.94 137.477 111.249C155.525 120.009 171.191 131.819 183.225 148.1C184.776 150.197 186.623 152.074 188.562 154.32Z",
-    },
-    {
-      id: "star",
-      viewBox: { width: 684, height: 690 },
-      faceCenter: { x: 325, y: 390 },
-      svgPath:
-        "M683.528 438.707C682.054 433.169 678.314 427.905 675.121 423.222C668.475 413.472 660.376 404.814 651.47 397.083C633.266 381.269 611.819 369.688 590.155 359.349C577.719 353.411 565.073 347.951 552.511 342.287C550.525 341.389 548.553 340.471 546.602 339.497C547.465 339.279 548.145 338.922 548.433 338.403C576.835 287.491 605.636 236.81 634.831 186.346C636.171 184.026 629.75 183.976 628.837 184.061C571.073 189.073 513.322 199.748 458.05 217.413C452.611 219.151 447.2 220.953 441.839 222.902C441.88 221.794 441.915 220.686 441.944 219.572C441.972 218.57 439.004 218.843 438.589 219.404C438.379 219.677 438.168 219.957 437.957 220.238C421.129 164.313 404.166 108.424 387.127 52.5623C381.962 35.6196 376.656 18.7188 371.533 1.76204C370.859 -0.481009 364.326 -0.537039 363.196 1.33474C342.507 35.4448 323.86 70.7745 307.46 107.134C299.263 125.311 291.599 143.726 284.596 162.386C277.795 180.499 271.949 198.928 265.38 217.119C264.033 220.834 262.629 224.508 261.198 228.159C254.734 223.554 248.278 218.948 241.815 214.343C227.351 204.025 212.88 193.713 198.409 183.394C169.039 162.456 139.571 141.672 110.194 120.74C93.4354 108.795 76.6835 96.8501 59.9249 84.9058C58.7598 84.0715 51.2506 83.2027 51.4891 85.782C54.4368 117.614 63.2369 148.422 72.4234 178.929C81.7782 210.018 90.5931 241.051 96.909 272.924C98.6565 281.749 100.235 290.61 101.639 299.499C102.081 302.296 102.502 305.086 102.979 307.869C97.6461 308.065 92.319 308.282 87.0067 308.808C62.0509 311.247 39.292 323.276 16.6947 333.293C13.7895 334.576 10.8837 335.873 7.97169 337.141C5.83086 338.073 3.4521 338.992 1.80996 340.73C-2.22514 344.999 1.40307 350.523 3.83793 354.568C9.95037 364.683 16.421 374.588 23.2281 384.248C37.0957 403.938 52.4508 422.578 69.0337 440.053C77.9885 449.495 87.2662 458.72 97.0143 467.356C101.31 471.156 106.187 474.723 110.629 478.649C105.703 483.907 101.015 489.395 96.4811 494.834C86.4803 506.835 77.2309 519.53 68.543 532.498C51.2722 558.288 36.4571 585.717 24.4776 614.346C21.5792 621.265 18.8422 628.246 16.2946 635.306C13.9435 641.811 10.7647 648.723 9.9436 655.634C9.26985 661.312 11.256 666.261 17.1859 667.712C25.4176 669.724 32.4146 664.95 39.1657 661.025C69.5676 643.331 99.66 625.148 129.409 606.375C155.277 590.049 180.893 573.31 206.241 556.192C214.361 570.1 223.217 583.615 232.263 596.898C245.92 616.953 260.454 636.434 275.62 655.375C284.357 666.289 293.319 677.021 302.477 687.585C303.895 689.218 310.541 689.913 311.671 687.585C333.756 642.106 351.504 594.55 364.74 545.769C409.697 579.837 456.274 611.767 504.305 641.377C518.559 650.167 532.946 658.74 547.444 667.124C549.444 668.28 554.757 668.918 556.364 666.682C562.497 658.144 561.367 647.503 558.56 637.935C555.044 625.942 549.865 614.333 545.381 602.676C535.576 577.195 525.78 551.721 515.975 526.247C510.698 512.536 505.427 498.824 500.15 485.12C529.906 482.856 559.521 478.972 588.856 473.498C604.127 470.645 619.307 467.351 634.381 463.628C641.814 461.791 649.217 459.85 656.586 457.803C663.127 455.987 670.201 454.585 676.334 451.62C681.535 449.095 685.135 444.75 683.528 438.707ZM216.052 549.828C215.182 551.103 214.045 552.106 213.104 553.297C212.732 553.767 212.431 554.293 212.136 554.825C211.729 554.223 211.322 553.62 210.922 553.017C212.739 551.783 214.564 550.549 216.381 549.309C216.276 549.484 216.171 549.659 216.052 549.828Z",
-    },
-    {
-      id: "ghost",
-      viewBox: { width: 594, height: 595 },
-      faceCenter: { x: 321, y: 167 },
-      svgPath:
-        "M576.611 203.334C566.222 162.978 549.573 124.108 524.337 90.7273C501.665 60.7321 472.541 35.223 438.426 19.0303C420.561 10.5482 401.503 4.96863 381.912 2.27672C357.828 -1.02474 333.415 -0.162803 309.225 1.28844C285.802 2.6972 261.921 3.77715 239.382 10.8708C217.944 17.6215 198.886 30.1551 181.232 43.7612C149.083 68.5339 120.5 97.9404 96.5148 130.648C72.3814 163.546 52.8605 199.836 38.9201 238.173C31.0023 259.952 25.0568 282.363 19.8062 304.921C14.6684 326.959 9.92941 349.118 6.34967 371.472C-0.550239 414.478 -3.40741 459.088 5.99142 501.954C8.22336 512.132 11.1649 522.156 14.8427 531.914C18.6611 542.043 23.0201 552.628 29.3448 561.481C35.3815 569.928 43.9666 574.947 53.604 578.389C64.3932 582.23 75.6244 585.097 86.8205 587.48C98.6412 589.997 110.623 591.848 122.662 592.976C132.987 593.951 143.467 594.637 153.821 593.551C163.087 592.576 172.528 590.116 180.208 584.656C187.711 579.307 192.654 571.512 195.693 562.932C198.865 553.973 200.143 544.482 201.21 535.081C202.446 524.09 203.737 513.127 205.604 502.226C209.549 479.199 215.432 456.417 224.094 434.692C231.955 414.974 241.768 394.715 256.39 379.048C269.734 364.755 289.993 353.161 310.06 356.911C324.724 364.797 334.278 380.03 339.388 395.487C345.839 415.009 345.18 435.387 342.891 455.575C340.23 479.016 335.289 502.85 338.005 526.522C339.15 536.518 342.105 546.508 348.534 554.436C353.974 561.137 361.359 565.925 369.263 569.247C387.24 576.797 407.758 577.463 426.97 576.846C450.471 576.089 473.825 572.247 496.28 565.28C515.739 559.244 534.853 550.412 549.537 535.992C564.629 521.167 573.438 501.315 579.355 481.316C592.938 435.415 594.195 386.731 593.1 339.226C592.053 293.572 588.025 247.671 576.611 203.334ZM334.25 371.977C333.541 371.031 332.804 370.105 332.046 369.201C334.264 371.612 336.195 374.283 337.774 377.143C336.672 375.37 335.5 373.652 334.25 371.977Z",
-    },
-    {
-      id: "urchin",
-      viewBox: { width: 794, height: 720 },
-      faceCenter: { x: 397, y: 338 },
-      svgPath:
-        "M294.515 186.33C290.926 182.311 287.411 178.223 283.736 174.285C255.181 143.686 237.726 106.858 224.624 67.7708C221.922 59.7082 221.014 50.6841 221.11 42.1235C221.295 25.7049 233.064 17.8357 248.541 23.4619C254.559 25.65 260.702 28.8342 265.378 33.1064C277.385 44.0774 289.777 54.9774 299.853 67.626C313.811 85.1474 325.792 104.243 338.617 122.665C340.383 125.204 342.137 127.751 345.66 129.793C345.259 121.324 344.546 112.855 344.526 104.384C344.49 88.8345 344.241 73.2395 345.329 57.7508C345.875 49.9765 348.303 42.0299 351.376 34.7965C359.52 15.637 376.269 13.3068 389.644 29.0276C397.8 38.6138 402.618 49.9948 405.881 61.9238C414.873 94.7912 421.728 128.112 426.003 161.94C426.402 165.096 426.969 168.231 429.108 171.932C430.344 169.342 431.737 166.813 432.795 164.152C451.865 116.184 478.991 73.1319 513.081 34.5321C520.917 25.6599 530.426 18.0812 539.948 10.9498C545.615 6.70501 552.535 3.58415 559.368 1.54792C576.16 -3.45791 588.665 3.8339 591.208 21.1556C593.08 33.9082 593.763 47.2524 592.171 59.9825C590.435 73.8528 586.186 87.5585 581.754 100.913C571.342 132.286 557.635 162.224 540.214 190.399C537.626 194.586 535.262 198.911 534.128 204.328C536.972 202.423 539.852 200.569 542.653 198.603C556.596 188.816 570.47 178.93 584.456 169.206C614.283 148.469 647.654 135.529 682.263 125.54C688.765 123.664 695.867 122.972 702.652 123.159C713.904 123.468 720.214 131.503 717.203 142.284C715.307 149.074 711.569 155.622 707.514 161.49C692.518 183.186 673.201 200.825 652.363 216.643C632.461 231.751 611.755 245.804 591.386 260.296C588.209 262.556 584.938 264.685 579.709 268.231C584.517 268.511 587.145 268.847 589.763 268.789C638.566 267.692 686.823 271.965 734.215 283.994C744.006 286.479 753.628 289.901 763.003 293.691C773.991 298.134 782.9 305.511 789.297 315.65C796.827 327.584 794.162 339.013 780.931 343.895C768.816 348.366 755.947 351.686 743.136 353.225C697.592 358.692 651.905 362.624 605.642 359.936C616.683 365.218 628.128 369.808 638.676 375.938C656.819 386.481 674.627 397.645 692.132 409.218C701.774 415.593 708.433 424.652 709.643 436.758C710.806 448.401 706.534 455.432 695.6 459.516C683.268 464.122 670.45 463.849 658.061 460.912C633.414 455.067 609.006 448.207 584.503 441.756C581.859 441.06 579.232 440.292 575.916 440.921C577.782 442.502 579.632 444.103 581.517 445.661C625.821 482.299 661.92 526.023 692.447 574.478C703.179 591.515 711.331 609.609 710.856 630.425C710.806 632.594 710.83 634.799 710.448 636.922C708.794 646.103 703.379 650.57 694.271 648.581C687.312 647.062 680.063 644.656 674.209 640.751C661.29 632.132 648.094 623.442 636.936 612.772C608.662 585.734 581.451 557.588 553.806 529.894C552.398 528.484 550.912 527.151 548.249 526.695C549.791 529.363 551.316 532.041 552.879 534.696C564.449 554.362 576.277 573.883 587.466 593.761C591.333 600.629 593.903 608.288 596.504 615.78C597.654 619.095 597.931 622.93 597.697 626.468C596.89 638.652 586.543 645.379 575.396 640.478C569.882 638.052 564.364 634.52 560.292 630.145C548.216 617.175 536.37 603.925 525.369 590.045C515.984 578.203 507.893 565.337 497.773 550.836C499.144 558.95 500.763 564.696 500.928 570.484C501.132 577.632 500.995 585.009 499.441 591.937C497.166 602.077 489.704 606.78 479.3 606.192C468.49 605.581 459.903 600.394 453.083 592.425C447.716 586.152 443.012 579.31 436.928 571.292C439.639 583.059 442.231 593.118 444.248 603.291C450.143 633.03 452.019 662.882 445.684 692.825C444.713 697.412 443.768 702.116 441.936 706.391C435.711 720.911 419.931 724.247 409.762 712.438C403.763 705.472 398.366 697.025 395.716 688.326C388.341 664.118 382.261 639.5 376.215 614.91C371.808 596.988 368.287 578.849 364.313 560.82C363.802 558.502 362.939 556.263 361.617 551.965C358.827 563.829 356.929 573.777 354.095 583.451C350.846 594.54 347.203 605.56 342.916 616.285C339.101 625.826 332.805 633.939 324.393 640.001C320.63 642.712 316.39 645.463 311.977 646.437C300.425 648.986 289.18 648.873 285.976 630.411C283.695 617.265 284.637 604.377 287.788 591.644C290.829 579.356 294.597 567.247 298.062 555.064C299.02 551.695 300.034 548.343 299.688 544.273C297.649 546.644 295.517 548.941 293.585 551.396C272.752 577.88 252.398 604.758 230.922 630.709C222.242 641.199 211.213 649.768 201.069 659.01C198.7 661.168 195.692 662.712 192.802 664.197C174.923 673.382 159.519 665.931 157.113 646.008C156.238 638.763 157.516 630.646 160.044 623.743C164.127 612.587 169.264 601.634 175.359 591.432C191.703 564.078 208.785 537.164 225.591 510.085C228.038 506.141 230.576 502.251 234.473 496.139C229.868 497.617 227.611 498.065 225.606 499.033C215.819 503.758 206.205 508.851 196.328 513.372C185.72 518.229 174.679 522.015 162.775 521.008C157.806 520.588 152.645 519.611 148.091 517.669C136.302 512.644 132.798 504.611 136.535 492.221C141.872 474.519 154.213 461.746 167.266 449.555C169.332 447.625 171.754 446.076 173.616 442.991C171.443 443.292 169.194 443.33 167.108 443.931C123.775 456.421 79.5185 459.784 34.6988 457.123C27.8868 456.718 20.9129 455.747 14.4336 453.706C0.908617 449.448 -3.50076 437.976 2.86467 425.249C6.70643 417.569 12.4396 411.581 20.2111 407.848C58.9649 389.23 98.9102 374.491 142.201 370.713C145.72 370.406 149.196 369.605 154.097 368.804C141.101 362.29 129.493 356.68 118.105 350.657C113.666 348.309 109.162 345.582 105.661 342.062C96.3634 332.713 98.382 319.883 110.258 314.244C116.706 311.183 123.841 308.878 130.896 307.92C145.554 305.93 160.366 305.065 175.791 303.676C173.747 302.193 171.727 300.55 169.543 299.169C150.261 286.972 130.969 274.793 111.642 262.669C98.6204 254.501 87.584 244.099 77.2847 232.8C68.5257 223.19 63.0141 211.886 59.6886 199.489C56.5094 187.638 61.3928 178.296 73.2079 175.124C88.4997 171.017 103.758 170.524 119.024 176.674C145.875 187.492 169.54 203.31 192.386 220.761C203.581 229.313 215.7 236.661 227.461 244.465C229.174 245.602 231.22 246.238 234.26 246.038C232.877 244.372 231.626 242.573 230.091 241.06C219.764 230.882 209.334 220.808 199.02 210.615C186.946 198.683 178.095 184.691 173.135 168.435C169.997 158.149 167.91 147.579 176.983 138.998C186.671 129.836 198.24 128.596 210.367 133.148C233.319 141.765 254.195 153.89 272.275 170.631C278.845 176.715 285.699 182.491 292.423 188.407C293.118 187.714 293.817 187.022 294.515 186.33Z",
-    },
-    {
-      id: "stack",
-      viewBox: { width: 632, height: 634 },
-      faceCenter: { x: 304, y: 358 },
-      svgPath:
-        "M82.907 607.208C99.1763 617.059 116.462 624.348 135.28 627.969C154.812 631.738 174.861 632.532 194.719 632.93C235.809 633.74 277.066 632.79 318.14 631.63C358.713 630.485 399.319 628.289 439.607 623.258C478.251 618.43 517.459 611.608 554.389 599.18C585.395 588.745 616.226 570.779 627.501 538.843C632.519 524.631 633.051 508.869 626.651 494.976C620.029 480.585 607.063 470.188 593.43 462.417C591.381 461.249 589.301 460.12 587.213 459.021C565.918 442.504 541.542 429.842 515.586 421.977C506.574 419.251 497.403 416.969 488.105 415.389C484.199 414.719 480.276 414.166 476.33 413.745C504.367 407.913 531.887 398.754 558.018 387.72C573.589 381.147 589.572 374.091 602.427 363.033C613.44 353.548 621.093 341.345 625 327.499C632.821 299.822 625.381 269.327 607.564 246.875C588.754 223.178 560.591 208.179 531.419 200.431C513.315 195.619 494.481 193.672 475.743 193.352C466.302 191.787 456.75 190.985 447.246 191.07C462.435 185.868 477.228 179.545 491.456 172.131C506.137 164.484 520.723 155.808 533.642 145.506C539.739 140.638 545.496 134.992 548.624 127.797C551.737 120.641 552.252 112.775 551.054 105.144C546.473 76.0653 523.899 50.8573 500.047 34.5815C472.654 15.8914 439.917 6.69447 407.084 2.78529C370.996 -1.52102 333.988 -0.399816 297.892 3.19781C261.257 6.85762 224.946 13.8355 189.621 24.0447C157.305 33.382 124.942 44.7906 95.674 61.3466C68.9717 76.4464 41.6732 98.0802 29.5806 126.738C23.8084 140.428 22.7443 155.084 27.7065 169.157C32.3117 182.2 41.0857 193.298 51.7492 202.144C71.2897 218.358 96.8411 225.234 122.297 227.305C72.7028 236.3 22.7283 259.001 4.98988 307.952C-4.26003 333.495 -0.941483 361.537 17.3925 382.416C35.1464 402.616 62.0632 413.535 88.186 418.69C98.3336 420.684 108.6 421.867 118.906 422.701C94.4429 429.678 70.8526 439.957 50.1685 454.294C22.6482 473.382 0.352613 506.081 12.7787 540.057C18.1385 554.729 28.786 566.722 40.4896 576.947C53.473 588.285 68.1388 598.26 82.907 607.208ZM486.287 418.931C486.35 418.915 486.533 418.97 486.398 419.001C483.405 419.71 480.419 419.211 477.846 417.56C477.767 417.513 478.013 417.475 478.069 417.513C480.538 419.103 483.405 419.609 486.287 418.931ZM448.945 193.772C444.681 193.982 440.425 194.263 436.177 194.637C437.456 194.248 438.734 193.858 440.012 193.453C442.998 193.499 445.975 193.609 448.945 193.772Z",
-    },
-    {
-      id: "flower",
-      viewBox: { width: 724, height: 691 },
-      faceCenter: { x: 335, y: 352 },
-      svgPath:
-        "M707.996 341.488C683.572 371.035 643.096 379.081 606.995 383.544C585.458 386.205 563.817 387.947 542.142 388.871C531.271 389.34 520.394 389.578 509.517 389.597C509.127 389.597 508.73 389.604 508.34 389.604C539.98 409.658 571.958 431.456 594.664 461.683C615.354 489.223 627.369 524.084 618.659 558.345C614.714 573.858 606.844 588.374 594.889 599.167C581.012 611.696 562.615 618.079 543.959 615.227C524.928 612.31 508.592 601.272 493.122 590.513C477.163 579.416 461.759 567.586 446.956 554.985C432.226 542.449 418.071 529.273 404.782 515.212C401.597 511.845 398.478 508.439 395.411 504.987C395.332 506.221 395.246 507.449 395.16 508.67C393.64 530.804 390.508 552.813 385.876 574.505C381.812 593.503 376.631 612.297 370.221 630.642C364.769 646.24 357.917 662.018 345.916 673.761C335.032 684.415 319.932 690.693 304.614 689.954C288.622 689.175 276.49 679.894 268.619 666.374C260.055 651.654 254.761 634.51 250.294 618.172C245.424 600.355 242.173 582.096 240.561 563.692C238.882 544.456 238.677 524.613 240.752 505.093C240.164 505.898 239.57 506.703 238.955 507.509C233.457 514.704 227.14 521.378 220.716 527.755C208.154 540.218 194.349 551.433 179.54 561.131C166.422 569.719 152.519 577.132 138.073 583.238C122.86 589.674 106.333 595.787 89.7861 597.351C74.8317 598.764 59.4075 595.437 49.4026 583.548C40.8116 573.336 37.4353 559.13 38.4529 546.02C39.6691 530.33 47.6584 515.879 57.1277 503.647C68.289 489.237 81.7962 476.568 94.8805 463.941C112.63 446.817 130.797 429.759 149.769 413.758C133.652 415.052 117.487 415.257 101.562 414.385C83.9244 413.414 65.996 411.269 49.079 406.015C34.3755 401.454 19.6257 393.756 10.2747 381.167C2.05413 370.11 -2.30039 355.053 1.24128 341.481C5.30564 325.922 19.0643 317.597 32.7163 311.115C64.555 296.005 99.8366 290.097 134.834 288.763C152.538 288.09 170.274 288.71 187.945 289.879C189.611 289.984 191.276 290.097 192.948 290.209C176.744 285.403 161.776 275.713 148.831 265.151C134.458 253.42 121.869 239.663 110.318 225.193C99.3547 211.456 88.8807 197.112 79.3448 182.351C69.558 167.188 61.3172 149.748 63.1485 131.237C66.4331 98.0067 102.368 85.2731 131.491 90.3755C149.836 93.5902 166.349 102.892 181.588 113.209C196.325 123.19 210.724 133.792 224.615 144.922C235.558 153.695 246.224 162.923 256.142 172.858C254.979 152.288 253.539 131.614 254.438 111.017C255.224 92.9499 258.066 74.7764 265.348 58.1084C272.082 42.6945 282.748 29.3267 296.143 19.1613C309.34 9.15344 326.309 1.12634 343.035 0.0771651C360.441 -1.01881 374.92 9.70203 384.813 23.1751C403.157 48.167 404.803 82.1373 400.666 111.764C397.904 131.527 392.564 150.843 388.461 170.35C386.386 180.225 384.575 190.166 383.524 200.207C383.445 200.973 383.379 201.738 383.313 202.511C384.529 200.299 385.863 198.121 387.284 195.949C392.887 187.394 399.146 179.149 405.674 171.274C418.838 155.398 433.225 140.565 448.688 126.907C474.956 103.711 506.371 78.8764 542.863 76.8298C551.982 76.3212 561.479 77.3051 569.778 81.319C576.968 84.7978 582.638 90.5806 586.583 97.4655C594.909 112.028 596.74 129.666 594.387 146.05C589.213 182.06 565.7 211.211 538.798 234.164C524.716 246.184 509.576 256.879 494.542 267.665C492.553 269.091 490.564 270.517 488.575 271.942C500.847 267.367 513.29 263.228 525.807 259.459C547.515 252.917 569.593 247.537 592.101 244.567C611.965 241.946 632.332 241.095 652.289 243.26C667.541 244.916 683.03 248.686 696.346 256.528C710.032 264.595 720.156 277.487 722.555 293.423C725.144 310.667 718.966 328.22 707.996 341.488Z",
-    },
-    {
-      id: "burst",
-      viewBox: { width: 734, height: 646 },
-      faceCenter: { x: 345, y: 305 },
-      svgPath:
-        "M733.742 331.374C731.601 337.333 725.454 340.564 719.925 342.794C712.107 345.948 703.946 348.402 695.982 351.156C687.464 354.1 678.952 357.044 670.432 359.982C662.068 362.877 653.717 365.8 645.36 368.709C646.518 374.625 647.563 380.563 648.49 386.521C650.343 398.515 651.718 410.579 652.595 422.678C654.377 447.114 654.125 471.663 651.928 496.064C651.71 498.475 645.999 498.026 644.574 497.676C620.435 491.696 596.169 486.229 571.799 481.308C560.613 479.051 549.4 476.913 538.172 474.88C537.828 490.687 536.418 506.53 534.404 522.049C531.182 546.843 525.807 571.336 518.369 595.211C517.457 598.127 511.225 597.209 509.261 596.157C490.35 586.021 471.487 575.78 452.681 565.441C443.208 560.232 433.671 555.094 424.332 549.654C421.006 547.719 417.652 545.792 414.36 543.78C412.248 546.507 409.406 548.96 407.287 551.28L396.445 563.169C388.901 571.441 381.358 579.712 373.822 587.984C360.307 602.81 346.791 617.629 333.283 632.448C327.992 638.252 323.241 645.598 314.638 645.801C305.93 646.012 302.33 639.794 298.92 632.826C290.225 615.056 281.489 597.308 272.907 579.481C267.16 567.544 261.469 555.564 256.592 543.261C253.743 543.674 250.866 543.969 248.115 544.39C241.736 545.371 235.351 546.325 228.958 547.257C216.643 549.052 204.307 550.762 191.963 552.388C166.765 555.704 141.51 558.473 116.213 560.989C115.182 561.094 110.255 561.136 111.083 559.026C117.925 541.704 126.248 525.035 136.064 509.207C140.942 501.335 146.268 493.722 151.909 486.383C153.039 484.911 154.19 483.432 155.347 481.953C154.239 481.735 153.123 481.518 152.007 481.315C140.464 479.261 128.83 477.726 117.174 476.485C93.9326 474.017 70.5584 472.524 47.2193 471.354C46.7841 471.333 44.0895 471.242 44.2368 470.212C47.0505 450.366 51.9067 430.9 58.7625 412.065C62.2498 402.482 66.2919 393.138 70.8388 384.011C70.9861 383.71 71.1476 383.415 71.2949 383.114C67.6738 381.166 64.0459 379.217 60.4322 377.261C52.2574 372.852 44.1104 368.401 35.9492 363.97C28.6373 360.002 21.3464 356.007 14.0696 351.976C10.3857 349.936 6.42068 348.065 2.9401 345.695C0.652791 344.146 -0.414131 342.036 0.147373 339.267C0.806877 336.043 2.86302 333.106 4.46309 330.288C6.06316 327.477 7.69098 324.694 9.36132 321.932C13.1156 315.714 17.0589 309.608 21.1855 303.629C29.4589 291.628 38.4688 280.118 48.2862 269.337C51.0789 266.267 53.9284 263.182 56.8475 260.14C55.29 258.598 53.8021 256.965 52.2933 255.416C48.3985 251.385 44.5813 247.284 40.848 243.113C34.2308 235.732 27.8662 228.119 21.7612 220.31C18.5684 216.23 15.4527 212.101 12.4141 207.909C10.1829 204.839 6.97595 201.089 7.19352 196.995C7.39014 193.224 10.6599 190.946 13.558 189.095C17.7684 186.403 22.0558 183.816 26.3154 181.188C35.2064 175.699 44.1387 170.266 53.1 164.89C71.3448 153.954 89.8283 143.384 108.466 133.128C112.01 131.18 115.553 129.21 119.104 127.254C119.048 126.49 119.048 125.656 119.104 124.71C119.258 122.011 119.504 119.319 119.736 116.627C120.234 110.795 120.81 104.97 121.476 99.1513C122.781 87.7464 124.409 76.3692 126.36 65.0554C128.079 55.0661 130.072 45.1265 132.332 35.2425C133.504 30.1111 134.647 24.9378 136.058 19.8699C136.9 16.8202 137.959 13.6382 140.527 11.5984C147.489 6.08155 157.516 10.4556 164.604 13.2738C172.835 16.5475 180.933 20.1574 188.876 24.09C204.945 32.0389 220.348 41.3202 234.88 51.807C241.848 56.8331 248.585 62.1535 255.076 67.7824C267.49 56.6644 280.268 46.0588 293.895 36.3224C311.887 23.4663 330.841 12.4961 349.689 0.999445C351.478 -0.0872931 357.857 -0.88646 358.628 1.9249C367.372 34.0018 375.014 66.6328 380.269 99.509C380.725 99.243 381.175 98.9832 381.624 98.7172C388.283 94.7637 395.118 91.0904 401.995 87.5155C450.806 62.0908 501.52 40.311 553.454 22.0785C554.395 21.749 562.577 20.9782 561.79 23.4035C555.474 42.9118 549.973 62.6726 545.342 82.6507C543.055 92.5274 540.978 102.447 539.118 112.414C538.164 117.511 537.265 122.614 536.423 127.724C536.051 129.975 535.546 132.266 535.399 134.552C535.336 135.562 535.273 136.746 535.575 137.734C535.946 138.001 536.185 138.246 536.298 138.484C537.378 138.204 539.097 138.232 539.469 138.982C539.476 138.982 539.476 138.989 539.476 138.996C545.708 137.615 552.058 136.641 558.198 135.659C570.513 133.682 582.941 132.316 595.375 131.404C617.641 129.764 639.983 129.68 662.256 131.145C673.624 131.888 684.971 133.052 696.255 134.636C701.728 135.407 707.201 136.256 712.647 137.23C716.408 137.896 721.208 138.905 722.695 142.935C724.141 146.868 721.369 151.172 719.236 154.284C715.958 159.058 712.639 163.79 709.341 168.549C702.745 178.061 696.149 187.567 689.56 197.08C677.273 214.794 664.992 232.507 652.713 250.214C652.004 251.707 651.19 252.948 650.025 254.097C650.011 254.112 649.998 254.132 649.99 254.147H649.976C649.822 254.294 649.66 254.448 649.491 254.595C649.365 254.707 649.176 254.806 648.937 254.897C653.071 257.603 657.063 260.533 661.014 263.464C668.515 269.016 675.785 274.883 682.936 280.883C695.434 291.376 707.489 302.396 719.025 313.927C722.168 317.075 725.277 320.25 728.344 323.467C730.212 325.423 734.78 328.479 733.742 331.374Z",
-    },
-    {
-      id: "ninja",
-      viewBox: { width: 789, height: 651 },
-      faceCenter: { x: 395, y: 387 },
-      svgPath:
-        "M785.439 391.136C742.323 396.551 699.565 404.289 657.437 414.312C622.285 422.682 586.81 432.073 554.578 447.8C526.696 461.411 502.19 479.985 481.115 501.547C439.044 544.612 410.503 597.153 382.908 648.718C381.906 650.583 374.482 650.615 373.488 648.718C359.865 622.528 345.697 596.462 328.397 572.131C313.374 550.989 295.878 530.989 274.51 515.003C251.481 497.771 224.488 486.122 195.296 481.12C164.347 475.815 132.325 477.881 101.529 483.057C69.4521 488.452 37.9419 496.906 7.02702 506.368C6.34046 506.581 -1.45434 506.716 0.240489 504.974C20.2641 484.425 39.6366 463.251 56.5788 440.443C84.9933 402.199 106.404 358.739 108.666 311.866C110.599 271.872 98.3426 229.471 67.8548 200.062C66.8114 199.984 65.8516 199.739 65.2846 199.203C64.5771 198.532 64.6962 197.978 65.1868 197.571C64.8368 197.254 64.5002 196.931 64.1501 196.622C63.4919 196.041 67.4277 195.415 68.3595 196.209C68.4926 196.325 68.6257 196.448 68.7588 196.564C88.0749 193.33 106.782 187.355 124.068 178.758C155.768 162.998 182.305 139.222 199.814 110.187C218.143 79.7965 226.407 44.6299 223.416 9.97993C223.193 7.35945 231.205 8.36674 232.487 9.11484C259.619 25.0237 286.822 40.9769 315.614 54.2523C365.482 77.2536 423.508 92.7228 479.041 79.493C525.175 68.5086 564.831 38.8213 588.868 1.26078C589.926 -0.391183 598.863 -0.598052 598.435 1.71889C591.648 39.0276 585.184 76.6656 585.443 114.607C585.66 146.618 591.137 178.512 604.528 208.173C615.889 233.336 632.117 256.402 650.956 277.332C690.065 320.778 739.415 355.189 788.35 388.607C790.6 390.142 786.447 391.007 785.439 391.136Z",
-    },
-  ],
-
-  eyeStyles: [
-    {
-      id: "grumpy",
-      sourceViewBox: { width: 580, height: 561 },
-      eyeCenter: { x: 300, y: 273 },
-      paths: [
-        {
-          svgPath:
-            "M408.462 318.767C309.849 318.767 308.699 226.908 308.699 226.908H508.224C508.224 226.908 512.033 318.767 408.462 318.767Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M491.195 226.908C489.95 244.612 480.356 297.536 410.187 297.536C343.374 297.536 331.544 244.612 329.459 226.908H491.195Z",
-          color: PUPIL,
-        },
-        {
-          svgPath:
-            "M190.364 318.767C288.977 318.767 290.126 226.908 290.126 226.908H90.602C90.602 226.908 86.7932 318.767 190.364 318.767Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M107.631 226.908C108.876 244.612 118.47 297.536 188.64 297.536C255.452 297.536 267.282 244.612 269.367 226.908H107.631Z",
-          color: PUPIL,
-        },
-      ],
-    },
-    {
-      id: "angry",
-      sourceViewBox: { width: 700, height: 639 },
-      eyeCenter: { x: 350, y: 380 },
-      paths: [
-        {
-          svgPath:
-            "M364.85 351.133L548.822 267C548.822 267 478.078 493.069 364.85 351.133Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M436.416 397.949C454.071 397.949 468.383 383.651 468.383 366.013C468.383 348.375 454.071 334.076 436.416 334.076C418.761 334.076 404.449 348.375 404.449 366.013C404.449 383.651 418.761 397.949 436.416 397.949Z",
-          color: PUPIL,
-        },
-        {
-          svgPath:
-            "M334.972 351.133L151 267C151 267 221.744 493.069 334.972 351.133Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M263.406 397.949C281.061 397.949 295.373 383.651 295.373 366.013C295.373 348.375 281.061 334.076 263.406 334.076C245.751 334.076 231.439 348.375 231.439 366.013C231.439 383.651 245.751 397.949 263.406 397.949Z",
-          color: PUPIL,
-        },
-      ],
-    },
-    {
-      id: "curious",
-      sourceViewBox: { width: 613, height: 628 },
-      eyeCenter: { x: 264, y: 415 },
-      paths: [
-        {
-          svgPath:
-            "M219.737 414.847C219.737 420.414 219.404 425.84 218.785 431.075C218.785 431.075 218.832 431.122 218.785 431.169C215.264 460.388 202.463 483.801 185.902 492.081C181.667 494.175 177.241 495.318 172.625 495.318C146.642 495.318 125.514 459.294 125.514 414.847C125.514 370.447 146.643 334.425 172.625 334.425C198.656 334.424 219.737 370.447 219.737 414.847Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M218.785 431.075C218.785 431.075 218.832 431.121 218.785 431.169C215.264 460.388 202.463 483.8 185.902 492.081C176.433 484.848 169.723 466.384 169.723 444.78C169.723 416.845 180.905 394.241 194.754 394.241C206.223 394.241 215.835 409.802 218.785 431.075Z",
-          color: PUPIL,
-        },
-        {
-          svgPath:
-            "M308.084 414.847C308.084 420.414 308.416 425.84 309.036 431.075C309.036 431.075 308.988 431.122 309.036 431.169C312.557 460.388 325.358 483.801 341.918 492.081C346.154 494.175 350.58 495.318 355.196 495.318C381.179 495.318 402.307 459.294 402.307 414.847C402.307 370.447 381.178 334.425 355.196 334.425C329.165 334.424 308.084 370.447 308.084 414.847Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M309.037 431.075C309.037 431.075 308.989 431.121 309.037 431.169C312.558 460.388 325.359 483.8 341.919 492.081C351.389 484.848 358.099 466.384 358.099 444.78C358.099 416.845 346.916 394.241 333.068 394.241C321.599 394.241 311.987 409.802 309.037 431.075Z",
-          color: PUPIL,
-        },
-      ],
-    },
-    {
-      id: "goofy",
-      sourceViewBox: { width: 684, height: 690 },
-      eyeCenter: { x: 325, y: 390 },
-      paths: [
-        {
-          svgPath:
-            "M307.139 391.203C307.139 436.064 289.447 474.171 264.846 487.926C258.491 491.485 251.673 493.412 244.578 493.412C210.028 493.412 182.018 447.652 182.018 391.203C182.018 334.754 210.028 288.994 244.578 288.994C251.673 288.994 258.491 290.921 264.846 294.48C289.447 308.235 307.139 346.342 307.139 391.203Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M307.138 391.203C307.138 436.064 289.447 474.171 264.846 487.926C240.238 474.177 222.547 436.064 222.547 391.203C222.547 346.342 240.238 308.229 264.846 294.48C289.447 308.236 307.138 346.342 307.138 391.203Z",
-          color: PUPIL,
-        },
-        {
-          svgPath:
-            "M467.078 388.777C467.078 433.638 449.387 471.745 424.785 485.5C418.43 489.059 411.613 490.986 404.518 490.986C369.967 490.986 341.957 445.226 341.957 388.777C341.957 332.328 369.967 286.568 404.518 286.568C411.613 286.568 418.43 288.496 424.785 292.054C449.386 305.809 467.078 343.916 467.078 388.777Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M467.078 388.777C467.078 433.638 449.386 471.745 424.785 485.5C400.178 471.751 382.486 433.638 382.486 388.777C382.486 343.916 400.178 305.804 424.785 292.055C449.386 305.81 467.078 343.916 467.078 388.777Z",
-          color: PUPIL,
-        },
-      ],
-    },
-    {
-      id: "surprised",
-      sourceViewBox: { width: 594, height: 595 },
-      eyeCenter: { x: 321, y: 167 },
-      paths: [
-        {
-          svgPath:
-            "M310.499 167.981C310.499 190.307 301.331 210.502 286.546 225.002C272.106 239.17 252.308 247.907 230.464 247.907C186.256 247.907 150.422 212.122 150.422 167.981C150.422 126.962 181.361 93.1593 221.203 88.5783C224.243 88.2274 227.332 88.0488 230.464 88.0488C274.665 88.0488 310.499 123.834 310.499 167.981Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M310.5 167.981C310.5 190.307 301.331 210.502 286.546 225.002C283.507 225.353 280.424 225.532 277.292 225.532C233.084 225.532 197.25 189.747 197.25 145.6C197.25 123.274 206.418 103.078 221.203 88.5783C224.243 88.2274 227.332 88.0488 230.464 88.0488C274.666 88.0488 310.5 123.834 310.5 167.981Z",
-          color: PUPIL,
-        },
-        {
-          svgPath:
-            "M491.382 164.755C491.382 187.081 482.214 207.277 467.429 221.777C452.989 235.944 433.191 244.681 411.347 244.681C367.139 244.681 331.305 208.896 331.305 164.755C331.305 123.737 362.244 89.9337 402.086 85.3527C405.125 85.0018 408.214 84.8232 411.347 84.8232C455.548 84.8232 491.382 120.609 491.382 164.755Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M491.382 164.755C491.382 187.081 482.214 207.277 467.429 221.777C464.389 222.128 461.306 222.306 458.174 222.306C413.967 222.306 378.133 186.521 378.133 142.374C378.133 120.048 387.301 99.8528 402.086 85.3527C405.126 85.0018 408.215 84.8232 411.347 84.8232C455.548 84.8232 491.382 120.609 491.382 164.755Z",
-          color: PUPIL,
-        },
-      ],
-    },
-    {
-      id: "bashful",
-      sourceViewBox: { width: 794, height: 720 },
-      eyeCenter: { x: 397, y: 338 },
-      paths: [
-        {
-          svgPath:
-            "M389.147 338.779C389.147 354.559 382.667 368.834 372.216 379.083C362.01 389.097 348.016 395.273 332.576 395.273C301.329 395.273 276 369.979 276 338.779C276 309.785 297.869 285.892 326.03 282.655C328.179 282.406 330.362 282.28 332.576 282.28C363.819 282.28 389.147 307.574 389.147 338.779Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M389.148 338.779C389.148 354.559 382.667 368.834 372.217 379.083C370.068 379.331 367.889 379.457 365.675 379.457C334.428 379.457 309.1 354.163 309.1 322.959C309.1 307.178 315.58 292.904 326.031 282.655C328.179 282.406 330.362 282.28 332.576 282.28C363.819 282.28 389.148 307.574 389.148 338.779Z",
-          color: PUPIL,
-        },
-        {
-          svgPath:
-            "M517.001 336.498C517.001 352.279 510.52 366.554 500.07 376.803C489.863 386.817 475.87 392.992 460.429 392.992C429.182 392.992 403.854 367.698 403.854 336.498C403.854 307.505 425.722 283.612 453.884 280.374C456.032 280.126 458.215 280 460.429 280C491.672 280 517.001 305.294 517.001 336.498Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M517.001 336.498C517.001 352.279 510.521 366.554 500.07 376.803C497.922 377.051 495.743 377.177 493.529 377.177C462.282 377.177 436.953 351.883 436.953 320.679C436.953 304.898 443.434 290.623 453.884 280.374C456.033 280.126 458.216 280 460.43 280C491.672 280 517.001 305.294 517.001 336.498Z",
-          color: PUPIL,
-        },
-      ],
-    },
-    {
-      id: "gentle",
-      sourceViewBox: { width: 632, height: 634 },
-      eyeCenter: { x: 304, y: 358 },
-      paths: [
-        {
-          svgPath:
-            "M229.473 469.065C258.726 469.065 282.441 421.52 282.441 362.87C282.441 304.22 258.726 256.675 229.473 256.675C200.219 256.675 176.504 304.22 176.504 362.87C176.504 421.52 200.219 469.065 229.473 469.065Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M229.473 417.424C244.502 417.424 256.685 392.999 256.685 362.869C256.685 332.739 244.502 308.314 229.473 308.314C214.445 308.314 202.262 332.739 202.262 362.869C202.262 392.999 214.445 417.424 229.473 417.424Z",
-          color: PUPIL,
-        },
-        {
-          svgPath:
-            "M376.988 459.719C406.242 459.719 429.957 412.174 429.957 353.524C429.957 294.874 406.242 247.329 376.988 247.329C347.734 247.329 324.02 294.874 324.02 353.524C324.02 412.174 347.734 459.719 376.988 459.719Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M376.989 408.079C392.017 408.079 404.2 383.653 404.2 353.524C404.2 323.394 392.017 298.969 376.989 298.969C361.96 298.969 349.777 323.394 349.777 353.524C349.777 383.653 361.96 408.079 376.989 408.079Z",
-          color: PUPIL,
-        },
-      ],
-    },
-    {
-      id: "quirky",
-      sourceViewBox: { width: 724, height: 691 },
-      eyeCenter: { x: 335, y: 352 },
-      paths: [
-        {
-          svgPath:
-            "M320.715 343.724C320.838 359.862 316.431 375.655 310.169 390.408C310.138 390.482 310.107 390.555 310.07 390.629C307.833 395.777 305.299 400.893 302.372 405.757C296.079 416.243 287.955 425.559 277.045 431.507C262.264 439.554 242.097 440.564 229.523 427.806C219.821 417.961 217.343 402.876 219.161 389.644C221.017 376.135 227.365 363.661 231.267 350.694C233.246 344.106 234.793 337.265 234.398 330.338C233.998 323.381 230.971 317.932 226.94 312.433C222.909 306.929 220.468 301.067 221.042 294.11C221.578 287.632 224.148 281.58 228.352 276.63C238.627 264.543 255.953 264.131 269.612 270.165C277.057 273.453 283.597 278.464 289.644 283.858C293.749 287.515 297.737 291.333 301.41 295.427C302.884 297.065 304.301 298.746 305.663 300.476C315.513 313.012 320.592 327.789 320.715 343.724Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M305.663 300.476C304.301 298.746 302.884 297.065 301.41 295.427C297.737 291.333 293.749 287.515 289.644 283.858C283.597 278.464 277.057 273.453 269.612 270.165C255.953 264.131 238.627 264.543 228.352 276.63C224.148 281.58 221.578 287.632 221.042 294.11C220.468 301.067 222.909 306.929 226.94 312.433C230.971 317.932 233.998 323.381 234.398 330.338C234.793 337.265 233.246 344.106 231.267 350.694C227.365 363.661 221.017 376.135 219.161 389.644C217.343 402.876 219.821 417.961 229.523 427.806C242.097 440.564 262.264 439.554 277.045 431.507C287.955 425.559 296.079 416.243 302.372 405.757C305.299 400.893 307.833 395.777 310.07 390.629C310.107 390.555 310.138 390.482 310.169 390.408C316.431 375.655 320.838 359.862 320.715 343.724C320.592 327.789 315.513 313.012 305.663 300.476ZM304.524 387.107C298.767 400.413 291.229 413.86 279.185 422.468C278.809 422.733 278.432 422.997 278.05 423.256C277.687 423.502 277.317 423.743 276.941 423.977C265.698 431.107 250.831 434.142 238.867 426.926C227.846 420.282 224.518 406.305 225.233 394.268C225.96 382.034 231.477 370.458 235.428 359.025C239.323 347.739 242.855 335.277 239.921 323.338C238.522 317.673 235.458 313.117 232.087 308.456C228.315 303.235 226.558 297.749 228.037 291.302C231.212 277.479 244.84 270.743 258.221 273.348C267.843 275.226 276.219 280.681 283.665 287.134C283.721 287.183 283.782 287.232 283.837 287.282C287.314 290.311 290.587 293.549 293.681 296.733C305.238 308.622 312.973 322.624 314.027 339.383C315.057 355.743 310.878 372.139 304.524 387.107Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M320.716 343.724C320.839 359.862 316.432 375.655 310.17 390.408C310.139 390.481 310.108 390.555 310.071 390.629C307.833 395.777 305.3 400.893 302.372 405.757C299.642 409.766 296.597 413.571 293.145 417.043C293.022 417.13 292.893 417.21 292.769 417.29C288.011 420.43 282.574 422.85 276.941 423.976C270.58 425.251 263.972 424.882 257.803 422.031C244.686 415.978 243.435 402.371 248.662 390.796C254.745 377.97 263.658 365.853 265.575 351.691C266.647 343.699 265.051 336.243 262.672 328.602C260.465 321.491 257.956 314.225 258.197 306.726C258.548 295.587 266.604 288.944 278.094 287.558C279.912 287.331 281.779 287.189 283.666 287.134C286.643 287.029 289.669 287.14 292.664 287.478C293.028 287.78 293.379 288.082 293.737 288.39C296.363 290.662 298.945 292.995 301.411 295.427C302.884 297.065 304.302 298.746 305.664 300.476C315.513 313.012 320.592 327.789 320.716 343.724Z",
-          color: PUPIL,
-        },
-        {
-          svgPath:
-            "M449.963 343.724C450.086 359.862 445.679 375.655 439.417 390.408C439.386 390.482 439.355 390.555 439.318 390.629C437.081 395.777 434.548 400.893 431.62 405.757C425.327 416.243 417.203 425.559 406.293 431.507C391.512 439.554 371.345 440.564 358.771 427.806C349.069 417.961 346.591 402.876 348.409 389.644C350.265 376.135 356.613 363.661 360.515 350.694C362.493 344.106 364.041 337.265 363.646 330.338C363.245 323.381 360.219 317.932 356.188 312.433C352.157 306.929 349.716 301.067 350.289 294.11C350.826 287.632 353.396 281.58 357.599 276.63C367.874 264.543 385.201 264.131 398.859 270.165C406.305 273.453 412.845 278.464 418.891 283.858C422.996 287.515 426.985 291.333 430.658 295.427C432.131 297.065 433.549 298.746 434.911 300.476C444.761 313.012 449.84 327.789 449.963 343.724Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M434.911 300.476C433.549 298.746 432.131 297.065 430.658 295.427C426.985 291.333 422.996 287.515 418.891 283.858C412.845 278.464 406.305 273.453 398.859 270.165C385.201 264.131 367.874 264.543 357.599 276.63C353.396 281.58 350.826 287.632 350.289 294.11C349.716 301.067 352.157 306.929 356.188 312.433C360.219 317.932 363.245 323.381 363.646 330.338C364.041 337.265 362.493 344.106 360.515 350.694C356.613 363.661 350.265 376.135 348.409 389.644C346.591 402.876 349.069 417.961 358.771 427.806C371.345 440.564 391.512 439.554 406.293 431.507C417.203 425.559 425.327 416.243 431.62 405.757C434.548 400.893 437.081 395.777 439.318 390.629C439.355 390.555 439.386 390.482 439.417 390.408C445.679 375.655 450.086 359.862 449.963 343.724C449.84 327.789 444.761 313.012 434.911 300.476ZM433.771 387.107C428.014 400.413 420.476 413.86 408.432 422.468C408.056 422.733 407.68 422.997 407.298 423.256C406.935 423.502 406.565 423.743 406.189 423.977C394.946 431.107 380.079 434.142 368.115 426.926C357.094 420.282 353.766 406.305 354.481 394.268C355.208 382.034 360.724 370.458 364.675 359.025C368.571 347.739 372.103 335.277 369.169 323.338C367.77 317.673 364.706 313.117 361.335 308.456C357.562 303.235 355.806 297.749 357.285 291.302C360.46 277.479 374.087 270.743 387.469 273.348C397.09 275.226 405.467 280.681 412.913 287.134C412.968 287.183 413.03 287.232 413.085 287.282C416.562 290.311 419.835 293.549 422.929 296.733C434.486 308.622 442.221 322.624 443.275 339.383C444.305 355.743 440.125 372.139 433.771 387.107Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M449.964 343.724C450.087 359.862 445.68 375.655 439.418 390.408C439.387 390.481 439.356 390.555 439.319 390.629C437.081 395.777 434.548 400.893 431.621 405.757C428.89 409.766 425.845 413.571 422.393 417.043C422.27 417.13 422.141 417.21 422.017 417.29C417.259 420.43 411.823 422.85 406.189 423.976C399.828 425.251 393.221 424.882 387.051 422.031C373.934 415.978 372.683 402.371 377.91 390.796C383.993 377.97 392.906 365.853 394.823 351.691C395.896 343.699 394.299 336.243 391.92 328.602C389.713 321.491 387.205 314.225 387.445 306.726C387.797 295.587 395.852 288.944 407.342 287.558C409.16 287.331 411.028 287.189 412.914 287.134C415.891 287.029 418.917 287.14 421.913 287.478C422.276 287.78 422.628 288.082 422.985 288.39C425.611 290.662 428.193 292.995 430.659 295.427C432.132 297.065 433.55 298.746 434.912 300.476C444.762 313.012 449.841 327.789 449.964 343.724Z",
-          color: PUPIL,
-        },
-      ],
-    },
-    {
-      id: "dazed",
-      sourceViewBox: { width: 734, height: 646 },
-      eyeCenter: { x: 345, y: 305 },
-      paths: [
-        {
-          svgPath:
-            "M313.693 304.831C313.693 328.312 303.574 349.436 287.451 364.087C273.217 377.028 254.291 384.918 233.522 384.918C189.241 384.918 153.352 349.06 153.352 304.831C153.352 260.597 189.241 224.744 233.522 224.744C254.291 224.744 273.217 232.634 287.451 245.569C303.574 260.22 313.693 281.344 313.693 304.831Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M313.692 304.831C313.692 328.311 303.573 349.435 287.45 364.086C277.647 369.564 266.343 372.69 254.313 372.69C216.795 372.69 186.383 342.31 186.383 304.831C186.383 267.352 216.795 236.966 254.313 236.966C266.343 236.966 277.647 240.092 287.45 245.569C303.573 260.22 313.692 281.344 313.692 304.831Z",
-          color: PUPIL,
-        },
-        {
-          svgPath:
-            "M375.883 304.831C375.883 328.312 386.002 349.436 402.124 364.087C416.359 377.028 435.284 384.918 456.053 384.918C500.334 384.918 536.224 349.06 536.224 304.831C536.224 260.597 500.334 224.744 456.053 224.744C435.284 224.744 416.359 232.634 402.124 245.569C386.002 260.22 375.883 281.344 375.883 304.831Z",
-          color: SCLERA,
-        },
-        {
-          svgPath:
-            "M375.883 304.831C375.883 328.311 386.002 349.435 402.124 364.086C411.927 369.564 423.231 372.69 435.262 372.69C472.779 372.69 503.192 342.31 503.192 304.831C503.192 267.352 472.779 236.966 435.262 236.966C423.232 236.966 411.927 240.092 402.124 245.569C386.002 260.22 375.883 281.344 375.883 304.831Z",
-          color: PUPIL,
-        },
-      ],
-    },
-  ],
-
-  // Palette lives in its own module so boot-path palette consumers don't
-  // statically depend on this file's SVG data.
-  colors: BUNDLED_COLORS,
-
-  faceCenterOverrides: [
-    // Ghost body — shift non-native eyes from y=167 -> y=200
-    {
-      bodyShape: "ghost",
-      eyeStyle: "grumpy",
-      faceCenter: { x: 321, y: 200 },
-    },
-    {
-      bodyShape: "ghost",
-      eyeStyle: "angry",
-      faceCenter: { x: 321, y: 200 },
-    },
-    {
-      bodyShape: "ghost",
-      eyeStyle: "curious",
-      faceCenter: { x: 321, y: 200 },
-    },
-    {
-      bodyShape: "ghost",
-      eyeStyle: "goofy",
-      faceCenter: { x: 321, y: 200 },
-    },
-    {
-      bodyShape: "ghost",
-      eyeStyle: "bashful",
-      faceCenter: { x: 321, y: 200 },
-    },
-    {
-      bodyShape: "ghost",
-      eyeStyle: "gentle",
-      faceCenter: { x: 321, y: 200 },
-    },
-    {
-      bodyShape: "ghost",
-      eyeStyle: "quirky",
-      faceCenter: { x: 321, y: 200 },
-    },
-    {
-      bodyShape: "ghost",
-      eyeStyle: "dazed",
-      faceCenter: { x: 321, y: 200 },
-    },
-    // Sprout body — shift non-native eyes from y=415 -> y=385
-    {
-      bodyShape: "sprout",
-      eyeStyle: "grumpy",
-      faceCenter: { x: 264, y: 385 },
-    },
-    {
-      bodyShape: "sprout",
-      eyeStyle: "angry",
-      faceCenter: { x: 264, y: 385 },
-    },
-    {
-      bodyShape: "sprout",
-      eyeStyle: "goofy",
-      faceCenter: { x: 264, y: 385 },
-    },
-    {
-      bodyShape: "sprout",
-      eyeStyle: "surprised",
-      faceCenter: { x: 264, y: 385 },
-    },
-    {
-      bodyShape: "sprout",
-      eyeStyle: "bashful",
-      faceCenter: { x: 264, y: 385 },
-    },
-    {
-      bodyShape: "sprout",
-      eyeStyle: "gentle",
-      faceCenter: { x: 264, y: 385 },
-    },
-    {
-      bodyShape: "sprout",
-      eyeStyle: "quirky",
-      faceCenter: { x: 264, y: 385 },
-    },
-    {
-      bodyShape: "sprout",
-      eyeStyle: "dazed",
-      faceCenter: { x: 264, y: 385 },
-    },
-  ],
-};
+export const BUNDLED_COMPONENTS: CharacterComponents = getCharacterComponents();
